Modernize invocation of NVIDIA driver installer
More recent versions of the NVIDIA driver installer have a
significantly changed interface, rendering the previous set of flags
employed here obsolete. Adjust the invocation of the installer to meet
its new interface, as well as the default URL template by which it can
be downloaded.

Note that we don't attempt to retain compatibility with the previous
installer interface. It's not possible to change just the
"NVIDIA_DRIVER_VERSION" variable back to, say, "396.26" and install
that version instead. Accommodating both of these known installer
interfaces (and any in between) would entail surveying each version's
installer to learn which changes in interfaces occurred when.