Adjust invocation of LTS NVIDIA driver installer (#6)
The "long-lived branch" NVIDIA driver's installer has changed its
interface yet again. Adjust the flags we pass to the installer,
retaining as much of the previous configuration as possible. (It
doesn't appear that there are any flags that allow omission of the
documentation being installed.)

Bump the default driver version to 410.78, which is the latest along
the "long-lived branch," promising "relatively long term stability."